[REF] Change type of the offset to off_t
When accessing storage we need to be able to address even large storages, so
the offset must be of an off_t type.

Fixes:
SEOS-1472

+++ include/lib_host/HostStorage.h
@@ -17,32 +17,32 @@
 #endif
 
 #if !defined(HOSTSTORAGE_SIZE)
-#   define HOSTSTORAGE_SIZE ((size_t)(1 * 1024 * 1024))
+#   define HOSTSTORAGE_SIZE (((off_t)1) * 1024 * 1024)
 #endif
 
 // Equivalent of CAmkES if_OS_Storage interface --------------------------------
 
 OS_Error_t
 HostStorage_write(
-    size_t  const offset,
+    off_t   const offset,
     size_t  const size,
     size_t* const written);
 
 OS_Error_t
 HostStorage_read(
-    size_t  const offset,
+    off_t   const offset,
     size_t  const size,
     size_t* const read);
 
 OS_Error_t
 HostStorage_erase(
-    size_t  const offset,
-    size_t  const size,
-    size_t* const erased);
+    off_t  const offset,
+    off_t  const size,
+    off_t* const erased);
 
 OS_Error_t
 HostStorage_getSize(
-    size_t* const size);
+    off_t*  const size);
 
 OS_Error_t
 HostStorage_getState(
